+++ knotx-server/src/main/java/io/knotx/server/configuration/KnotxServerConfiguration.java
@@ -34,8 +34,13 @@ public class KnotxServerConfiguration {
 
   private JsonObject serverOptions;
 
+  private JsonObject customResponseHeader;
+
   public KnotxServerConfiguration(JsonObject config) {
     displayExceptionDetails = config.getBoolean("displayExceptionDetails", false);
+
+    customResponseHeader = config
+        .getJsonObject("customResponseHeader", new JsonObject());
 
     allowedResponseHeaders = config.getJsonArray("allowedResponseHeaders").stream()
         .map(item -> ((String) item).toLowerCase())
@@ -50,6 +55,10 @@ public class KnotxServerConfiguration {
 
   public boolean displayExceptionDetails() {
     return displayExceptionDetails;
+  }
+
+  public JsonObject getCustomResponseHeader() {
+    return customResponseHeader;
   }
 
   public Set<String> getAllowedResponseHeaders() {